+++ tests/_test_coro_vdecorator.py
@@ -1,7 +1,11 @@
 from asyncio import get_event_loop, coroutine
 
-from covador.vdecorator import ValidationDecorator
+from covador.vdecorator import ValidationDecorator, mergeof
 from covador import schema
+
+
+def async_run(coro):
+    return get_event_loop().run_until_complete(coro)
 
 
 def test_coro_func():
@@ -13,8 +17,7 @@ def test_coro_func():
     def boo(_, arg):
         return arg
 
-    loop = get_event_loop()
-    res = loop.run_until_complete(boo({'arg': 1}))
+    res = async_run(boo({'arg': 1}))
     assert res == 1
 
 
@@ -26,8 +29,7 @@ def test_coro_getter():
     def boo(_, arg):
         return arg
 
-    loop = get_event_loop()
-    res = loop.run_until_complete(boo({'arg': 1}))
+    res = async_run(boo({'arg': 1}))
     assert res == 1
 
 
@@ -40,6 +42,23 @@ def test_coro_func_getter():
     def boo(_, arg):
         return arg
 
-    loop = get_event_loop()
-    res = loop.run_until_complete(boo({'arg': 1}))
+    res = async_run(boo({'arg': 1}))
     assert res == 1
+
+
+def test_coro_validator_merge():
+    def getter0(it): return it[0]
+
+    @coroutine
+    def getter1(it): return it[1]
+
+    v0 = ValidationDecorator(getter0, None, schema)
+    v1 = ValidationDecorator(getter1, None, schema)
+    v = mergeof(v0, v1)
+
+    @v(arg=int)
+    def boo(_, arg):
+        return arg
+
+    res = async_run(boo([{}, {'arg': 3}]))
+    assert res == 3